feat: complete c parser (#13)
* fix: fix hardcode in export.go

* feat: cxx/c parser skeleton with clangd-18

* feat: initial support for cxx collect

1. Since clangd does not support semanticTokens/range method, use
  semanticTokens/full + filtering to emulate.
2. Since the concept of package and module does not apply to C/C++,
  treat the whole repo as a single package/module.

* fix: get json.Types and build graph

* fix: fix comments

* test: duplicate c functions

* refactor: clang-format c tests

* fix: pkgpath -> file. skip build.

C allows symbols with the same name in a single module, provided either:
* One is a weak symbol (decl) and one is a strong symbol (def)
* They are both strong symbols, but never linked together.

The first one works fine, but more changes are needed for the second
one.

testdata/cxxsimple illustrates the first scenario. Two instances of
`myself` are present, one (weak) in `pair.h` and one (strong) in `pair.c`.
The dependency is well defined in this scneario:
1. `pair.c:myself` depends on `pair.h:myself`
2. any other function using `myself` depends on both.
To verify, run `./abcoder parse cxx testdata/cxxsimple > cxxsimple.json`.

testdata/cxxduplicate is the second scenario. Two strong instances of
`add` are present, each used in a different executable. clangd handles
this with compile_commands.json. If clangd is invoked as below, the
`main->add` dependency shall point to the `add` in `d1/add.c`.

	mkdir build && cd build && cmake ..
	bear -- make prog1 # generate compile_commands.json
	cd testdata/cduplicate && clangd-18

While clangd does the right job, the current implementation of scanning
during collection does not take into account which files are included in
a compilation (as specified in compile_commands.json). So
`Collector.Collect` will incorrectly include `d2/add.c` even if it is
not used, and mess up with dependencies. That is to say, even for the
compilation `prog1 <- main.c, d1/add.c`, a dependency
`main->d2/add.c:add` will be present.